requirements.txt
yfinance
pandas
numpy
nltk
scrapy
beautifulsoup4
scikit-learn
plotly
streamlit
newsapi-python